Fix bug where iPad was not selected when it was turned on
When the iPad was turned on, there were two entries with the
same name for the iPad. One to connect mouse and keyboard and
one to start sidecar. Since the one to connect mouse and
keyboard comes first, this option was selected.

With this change, the script selects the last entry
with a matching name.